move place of the drawing function
