+++ pgdrive/scene_creator/map.py
@@ -2,12 +2,11 @@ import copy
 import json
 import logging
 import os
-from typing import List, Optional, Union, Iterable
+from typing import List
 
 import numpy as np
 from panda3d.core import NodePath
 
-from pgdrive.constants import Decoration
 from pgdrive.pg_config import PGConfig
 from pgdrive.pg_config.pg_blocks import PGBlock
 from pgdrive.scene_creator.algorithm.BIG import BIG, BigGenerateMethod
@@ -203,80 +202,6 @@ class Map:
         describe = self.random_seed if self.random_seed is not None else "custom"
         logging.debug("Scene {} is destroyed".format(describe))
 
-    # ==========
-    # The below code is deprecated. Part of it's function is inherited in top_down_observation.
-    # The code is used to draw the overall road network and annotate a given route with color.
-    #
-    def draw_maximum_surface(self, simple_draw=True) -> pygame.Surface:
-        import logging
-        from pgdrive.world.top_down_observation.top_down_obs_impl import WorldSurface, LaneGraphics
-        logging.warning("This part is deprecated! We will remove the rendering function in map instance future!")
-        surface = WorldSurface(self.film_size, 0, pygame.Surface(self.film_size))
-        b_box = self.road_network.get_bounding_box()
-        x_len = b_box[1] - b_box[0]
-        y_len = b_box[3] - b_box[2]
-        max_len = max(x_len, y_len)
-        # scaling and center can be easily found by bounding box
-        scaling = self.film_size[1] / max_len - 0.1
-        surface.scaling = scaling
-        centering_pos = ((b_box[0] + b_box[1]) / 2, (b_box[2] + b_box[3]) / 2)
-        surface.move_display_window_to(centering_pos)
-        for _from in self.road_network.graph.keys():
-            decoration = True if _from == Decoration.start else False
-            for _to in self.road_network.graph[_from].keys():
-                for l in self.road_network.graph[_from][_to]:
-                    if simple_draw:
-                        LaneGraphics.simple_draw(l, surface)
-                    else:
-                        two_side = True if l is self.road_network.graph[_from][_to][-1] or decoration else False
-                        LaneGraphics.display(l, surface, two_side)
-        return surface
-
-    def get_map_image_array(self, resolution: Iterable = (512, 512)) -> Optional[Union[np.ndarray, pygame.Surface]]:
-        import cv2
-        surface = self.draw_maximum_surface()
-        ret = cv2.resize(pygame.surfarray.pixels_red(surface), resolution, interpolation=cv2.INTER_LINEAR)
-        return ret
-
     @property
     def num_blocks(self):
         return len(self.blocks)
-
-    # def draw_navi_line(self, vehicle, dest_resolution=(512, 512), save=False, navi_line_color=(255, 0, 0)):
-    #
-    #     raise ValueError("This part is deprecated!")
-    #     """
-    #     Use this function to draw the whole map, with a navigation line.
-    #     :param vehicle: ego vehicle, BaseVehicle Class
-    #     :param dest_resolution: image resolution
-    #     :param save: save this image
-    #     :param navi_line_color: color of navigation line
-    #     :return: pygame.Surface
-    #     """
-    #     self.film_size = (2 * dest_resolution[0], 2 * dest_resolution[1])
-    #     checkpoints = vehicle.routing_localization.checkpoints
-    #     max_surface = self.draw_maximum_surface(simple_draw=False)
-    #     map_surface = pygame.Surface(dest_resolution)
-    #     pygame.transform.scale(max_surface, dest_resolution, map_surface)
-    #     surface = WorldSurface(self.film_size, 0, pygame.Surface(self.film_size))
-    #     b_box = self.road_network.get_bounding_box()
-    #     x_len = b_box[1] - b_box[0]
-    #     y_len = b_box[3] - b_box[2]
-    #     max_len = max(x_len, y_len)
-    #     # scaling and center can be easily found by bounding box
-    #     scaling = self.film_size[1] / max_len - 0.1
-    #     surface.scaling = scaling
-    #     centering_pos = ((b_box[0] + b_box[1]) / 2, (b_box[2] + b_box[3]) / 2)
-    #     surface.move_display_window_to(centering_pos)
-    #     for i, c in enumerate(checkpoints[:-1]):
-    #         lanes = self.road_network.graph[c][checkpoints[i + 1]]
-    #         for lane in lanes:
-    #             LaneGraphics.simple_draw(lane, surface, color=navi_line_color)
-    #     dest_surface = pygame.Surface(dest_resolution)
-    #     pygame.transform.scale(surface, dest_resolution, dest_surface)
-    #     dest_surface.set_alpha(100)
-    #     map_surface.blit(dest_surface, (0, 0))
-    #     if save:
-    #         pygame.image.save(map_surface, "map_{}.png".format(self.random_seed))
-    #     self.film_size = (self.config["draw_map_resolution"], self.config["draw_map_resolution"])
-    #     return map_surface
